smenovac/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
gunicorn>=21.0.0
waitress>=3.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
Pillow>=10.0.0
